pyproject.toml
[project]
name = "rag-byok-backend"
version = "0.1.0"
description = "A simple RAG client for personal use."
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "boto3>=1.42.30",
    "fastapi>=0.126.0",
    "huggingface>=0.0.1",
    "langchain>=1.0.7",
    "langchain-anthropic>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph[postgres]>=1.0.3",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.6",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "supabase>=2.27.0",
    "tiktoken>=0.12.0",
    "langgraph-checkpoint-postgres>=3.0.3",
    "langchain-openai>=1.1.7",
    "langchain-google-vertexai>=3.2.1",
    "psycopg-pool>=3.3.0",
]
